% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymdim-package.R
\docType{package}
\name{pymdim-package}
\alias{pymdim}
\alias{pymdim-package}
\title{pymdim: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/pymdim}
  \item Report bugs at \url{https://github.com/mdsumner/pymdim/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

}
\keyword{internal}
